pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-agent-platform"
version = "0.1.0"
description = "Multi-agent AI application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    # LLM providers
    "google-genai>=1.0.0",
    # Vector databases
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    # Tools
    "tavily-python>=0.5.0",
    "asyncpraw>=7.7.0",
    "exa-py>=1.0.0",
    # Observability
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "adapters", "apps"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
